fix: 테스트 클래스가 도메인 클래스명과 다른 문제 해결
